Introduced TrimCircularPitch
This was added by Ben Fraser, I believe in an effort to make the gears
mesh better. This would normally have been done with the clearance
and/or backlash parameters, but I didn't want to mess with success
during the Clockathon series. I may roll this one back in the future…