Workaround in drum() module for difference() bug in openscad
There is an inconsistency in openscad in how difference works on Macs and Linux.
When the drumHeight is 0, i.e. escapementWheel(), then the first term in the difference() for
the drum() module will be empty.  This causes the Mac to treat the second term as the actual
object and render it.  This leaves an artifact on the escapement wheel that is a solid disk which must
be drilled out manually to recover the axis hole for the escapement wheel. On linux platforms this
artifact doesn't appear.

For an openscad example that demonstrates the problem see:
 https://github.com/openscad/openscad/issues/221